requirements.txt
openai>=1.0.0
numpy>=1.21.0
tqdm>=4.64.0
requests>=2.28.0
matplotlib>=3.5.0
pandas>=1.3.0
python-dotenv>=0.19.0
datasets>=2.0.0 